features: - All quantities are now integers - Used dish_id instead of storing data about dish in cart - No loop for deleting items in cart
Really stupid mistakes

+++ dishes/models.py
@@ -7,7 +7,7 @@ class Dish(Document):
     price: Indexed(float)
     category: str
     description: str
-    quantity: Indexed(float)
+    quantity: Indexed(int)
 
     class Settings:
         name = "dishes"
